align to Megatron ref (scaling_moe_00196) at bf16 ULP level over 7485 steps
Bugs fixed:
- use_sequential string match: dataset='cybertron_baseline' (ends _baseline, not
  _cybertron) → training on torch.randint random samples, not Megatron-aligned
  sequential data. Match on `'cybertron' in dataset` instead.
- resume path didn't propagate non-arch keys (moe_routing_type, use_eod_attn_mask,
  eod_token_id, mask_loss_id, seq_aux_balance_alpha) → silent fallback to defaults.
  Flow through globals().get() from current config.
- LR off-by-one: seed ckpt stored iter_num=Adam step=1 (from PAI zero-grad
  patch), first training step got (1+1)*64/32000*peak = 2x ref's LR. Use
  Megatron's `iteration` field (completed training steps) instead.
- DDP find_unused_parameters=True unnecessary (grouped_mm touches all experts);
  flip to False, reclaims some MFU.
- Adam optim state must be loaded from Megatron distrib_optim.pt — recovery
  script reads 4 EP shards, bucket-slices exp_avg/exp_avg_sq/step with reverse
  iteration order verified via param value match.
- MoE aux-free bias update: was per-forward, should be per-optim-step
  (finalize_model_grads pattern). Split into local_tokens_per_expert accumulator
  + update_expert_bias() hook called before optimizer.step().
- Loss masks (eod_mask_loss, mask_loss_id=160000) + seq_aux alpha=0.0001.
- moe_routing_type='greedy' (flat top-K sigmoid, not group_limited_greedy).
- uint16 → int32 for token storage (Qwen vocab 152064 > 65535 wraps).
- Sample stride: prepare writes stride=block_size, train reads [s+1, s+block_size+1].
- megatron_to_nano.py: QKV interleaved GQA split, SwiGLU [gate;up] split,
  144 experts across 4 EP shards → stacked tensors.

Result (7485 steps DDP 8xH100):
- iter 0 bitwise aligned with ref iter 1 (loss Δ = +0.0001)
- 20-iter rolling avg Δ stays within ±0.01 nat through full run
- final 50-iter avg Δ = +0.003 nat (3 mnat, bf16 ULP ceiling)

Residual: mid-training (iter 2000-7000) nano ~0.006-0.009 nat higher than ref,
direction systematic. Likely bf16 kernel path differences (PyTorch SDPA vs TE
flash-attn-2, dense MoE dispatch vs allgather, PyTorch fused AdamW vs apex
FusedAdam). See ALIGNMENT.md for full breakdown.

+++ prepare_cybertron_data.py
@@ -153,7 +153,7 @@ def get_sample_tokens_cached(indexed_ds, doc_idx, sample_idx, actual_sample_id, 
     else:
         tokens = tokens[:seq_len]
 
-    return tokens.astype(np.uint16)
+    return tokens.astype(np.int32)
 
 
 def _extract_dataset_worker(args):
@@ -190,7 +190,7 @@ def _extract_dataset_worker(args):
     pairs.sort()
 
     # Open output memmap for writing
-    out = np.memmap(out_path, dtype=np.uint16, mode='r+', shape=(n_total * SEQ_LENGTH,))
+    out = np.memmap(out_path, dtype=np.int32, mode='r+', shape=(n_total * SEQ_LENGTH,))
 
     for actual_id, i in pairs:
         tokens = get_sample_tokens_cached(indexed_ds, doc_idx, sample_idx, actual_id, SEQ_LENGTH)
@@ -242,7 +242,7 @@ def extract_train_data(n_samples, out_path):
 
     # Write output
     os.makedirs(os.path.dirname(out_path), exist_ok=True)
-    out = np.memmap(out_path, dtype=np.uint16, mode='w+', shape=(n_samples * SEQ_LENGTH,))
+    out = np.memmap(out_path, dtype=np.int32, mode='w+', shape=(n_samples * SEQ_LENGTH,))
     print(f"Writing {n_samples:,} samples → {out_path} ({n_samples * SEQ_LENGTH * 2 / 1e9:.2f} GB)")
 
     import multiprocessing as mp
@@ -288,21 +288,21 @@ def extract_val_data(out_path, n_samples=2000):
         print(f"WARNING: Validation data not found at {VAL_DATA_PATH}")
         print("Creating minimal placeholder...")
         os.makedirs(os.path.dirname(out_path), exist_ok=True)
-        np.zeros(n_samples * SEQ_LENGTH, dtype=np.uint16).tofile(out_path)
+        np.zeros(n_samples * SEQ_LENGTH, dtype=np.int32).tofile(out_path)
         return
 
     indexed_ds = open_indexed_dataset(VAL_DATA_PATH)
     print(f"  Validation dataset: {len(indexed_ds)} docs")
 
     os.makedirs(os.path.dirname(out_path), exist_ok=True)
-    out = np.memmap(out_path, dtype=np.uint16, mode='w+', shape=(n_samples * SEQ_LENGTH,))
+    out = np.memmap(out_path, dtype=np.int32, mode='w+', shape=(n_samples * SEQ_LENGTH,))
 
     written = 0
     doc_idx = 0
-    buf = np.array([], dtype=np.uint16)
+    buf = np.array([], dtype=np.int32)
 
     while written < n_samples and doc_idx < len(indexed_ds):
-        doc = indexed_ds.get(doc_idx).astype(np.uint16)
+        doc = indexed_ds.get(doc_idx).astype(np.int32)
         buf = np.concatenate([buf, doc])
         doc_idx += 1
